Inventory & Catalog modules perf: replace entities fields 'id' by super class field 'id'. perf: remove extra class casts at the AbstractDao. perf: add singleton pattern for EntityManager.

+++ inventory/src/main/java/com/netcracker/sharlan/bean/OrderStatus.java
@@ -3,13 +3,13 @@ package com.netcracker.sharlan.bean;
 public enum OrderStatus {
     NEW(0), PENDING(1), DELIVERY(2), CLOSED(3), CANCELED(10);
 
-    private int id;
+    private long id;
 
-    OrderStatus(int id){
+    OrderStatus(long id){
         this.id = id;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
